+++ src/quicksort.c
@@ -54,7 +54,7 @@ static int qsort_cmp_wrap(void *ctx, const void *a, const void *b)
 #elif(!defined HAVE_QSORT_R)
 
 static inline char    *med3(char *, char *, char *, sort_cmp_t *, void *);
-static inline void     swapfunc(char *, char *, int, int);
+static inline void     swapfunc(char *, char *, size_t, int);
 
 #define min(a, b)    (a) < (b) ? a : b
 
@@ -62,7 +62,7 @@ static inline void     swapfunc(char *, char *, int, int);
  * Qsort routine from Bentley & McIlroy's "Engineering a Sort Function".
  */
 #define swapcode(TYPE, parmi, parmj, n) {         \
-    long i = (n) / sizeof (TYPE);             \
+    size_t i = (n) / sizeof (TYPE);             \
     TYPE *pi = (TYPE *) (parmi);         \
     TYPE *pj = (TYPE *) (parmj);         \
     do {                         \
@@ -76,7 +76,7 @@ static inline void     swapfunc(char *, char *, int, int);
     es % sizeof(long) ? 2 : es == sizeof(long)? 0 : 1;
 
 static inline void
-swapfunc(char *a, char *b, int n, int swaptype)
+swapfunc(char *a, char *b, size_t n, int swaptype)
 {
     if(swaptype <= 1)
         swapcode(long, a, b, n)
